Prevent from purchasing past events
As discussed in the topic below, although this is not currently the best
implementation, since it does not yet hides the option from the frontend
user, it blocks from adding a past event to the cart.

This works with a new configuration option in the 'Inventory' tab of the
'Edit Product' in the admin page. Once enabled, a number field will
appear, and there you can choose how long before the registration is
considered closed. If let at 0, purchases made on the same day will not
be allowed. A higher number will prevent X days before.

This commit has 3 unsolved issues:
- it does not take into account timezones. The server time is the time
that's being used. This has pluses and negatives. Security is the
positive side, since no one can pretend to change timezone to exploit,
but the negative side is that the user may not enjoy it. I honestly
prefer this way so we don't need to deal with security.

- the wc_add_notice for some reason is being called 3, 4, 5 and even 6
times after a failed attemp purchase. I do not know why, this is
probably due to the action being registred on the class? I thought the
init() was static and only ran once.

- the pt_BR translations, although they are present, they do not seem to
be working and I couldn't track down the issue.

https://wordpress.org/support/topic/hide-past-dates/