Fixed reports for latest version
First, this fixes that the reports wouldn't display if the variation wasn't enabled anymore.

Then, it was really inneficient, fetching a whole lot of data, just to join on the PHP, instead of just fetching what was needed from the DB.

Now it's able to parse all variations from registrations, including the ones that were already closed, since it's often that the event is closed a few hours or days before it actually starts.

Also, it waas not parsing older orders, so busy stores with lots of orders and events were just left out.

It's an improvement in all areas on my eyes.